+++ Calculadora/Commands/CalculatorCommands/ClearDisplayCommand.cs
@@ -1,18 +1,19 @@
 ﻿using Calculadora.Models;
 using Calculadora.ViewModels;
 
-namespace Calculadora.Commands
+namespace Calculadora.Commands.CalculatorCommands
 {
-    public class PercentageCommand : BaseCommand
+    public class ClearDisplayCommand : BaseCommand
     {
         private readonly StandardCalculatorViewModel _viewModel;
-        public PercentageCommand(StandardCalculatorViewModel viewModel)
+        public ClearDisplayCommand(StandardCalculatorViewModel viewModel)
         {
             _viewModel = viewModel;
         }
+
         public override void Execute(object? parameter)
         {
-            CalculatorDisplay.InsertPercentageInDisplay();
+            CalculatorDisplay.ClearDisplay();
             _viewModel.UpdateDisplay();
         }
     }
