+++ Calculadora/Commands/CalculatorCommands/LoadHistoryCommand.cs
@@ -4,7 +4,7 @@ using Calculadora.ViewModels;
 using System.Threading.Tasks;
 using System.Windows.Controls;
 
-namespace Calculadora.Commands
+namespace Calculadora.Commands.CalculatorCommands
 {
     public class LoadHistoryCommand : BaseAsyncCommand
     {
